layout/base/tests/test_frame_reconstruction_for_svg_transforms.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1419764
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1419764</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1419764 **/

  SimpleTest.waitForExplicitFinish();

  function run() {
    var utils = SpecialPowers.getDOMWindowUtils(window);
    var rect = document.querySelector("rect");

    var matrix = rect.transform.baseVal[0].matrix;

    matrix.e = 100;
    document.documentElement.offsetTop; // flush layout

    var startcount = utils.framesConstructed;

    matrix.e = 200;
    document.documentElement.offsetTop; // flush layout

    var endcount = utils.framesConstructed;
    is(endcount, startcount, "should not do frame construction");
    SimpleTest.finish();
  }

  </script>
</head>
<body onload="run()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1419764">Mozilla Bug 1419764</a>
<svg>
  <rect transform="translate(1 1)" width="20" height="20" fill="yellow"/>
</svg>
<pre id="test">
</pre>
</body>
</html>
